SPT: Isolate window globals
